fix data lossing.

+++ cpp/src/file/tsfile_io_reader.cc
@@ -377,8 +377,8 @@ int TsFileIOReader::load_all_measurement_index_entry(
   return ret;
 }
 
-int TsFileIOReader::read_device_meta_index(int32_t start_offset,
-                                           int32_t end_offset,
+int TsFileIOReader::read_device_meta_index(int64_t start_offset,
+                                           int64_t end_offset,
                                            common::PageArena &pa,
                                            MetaIndexNode *&device_meta_index,
                                            bool leaf) {
